Extract FolderTreeBuilder to UI/Helpers
Move BuildTree/BuildFolderNode recursive tree construction into a
static FolderTreeBuilder class. Code-behind BuildTree becomes a
3-line wrapper that delegates and sets ItemsSource.
